Create split.py
This specific sub - program takes all the pages in the document and distinguishes two following pages belonging to each other though keeping the difference between first and last page and then sends it to another directory. By changing the for loop you can change the order by which pages are separated.